Add files via upload
This MATLAB code is derived from a CellML implementation of the O'Hara-Rudy Model, with modifications to facilitate the generation of cardiac action potentials. It enables the simulation of electrical activity of cardiac cells based on the O'Hara-Rudy framework.

For a more detailed understanding, refer to "Simulation of the Undiseased Human Cardiac Ventricular Action Potential: Model Formulation and Experimental Validation" by 
Thomas O'Hara, László Virág, András Varró and Yoram Rudy.